// src/lib/hooks/types/assetStatusDetail.ts
export interface AssetStatusDetail {
  linkSerialNo: string;
  assetNo: string | null;
  parentAssetNo: string | null;
  accountingStandard: string | null;
  accountingDivisionCode: string | null;
  assetCategoryCode: string | null;
  assetName: string | null;
  detailBusinessName: string | null;
  detailBusinessCode: string | null;
  legalDongCode: string | null;
  detailAddress: string | null;
  acquisitionDept: string | null;
  managementDept: string | null;
  publicPropertyItemType: string | null;
  firstAcquisitionDate: Date | null;
  acquisitionDate: Date | null;
  acquisitionType: string | null;
  assetCurrentStatusCode: string | null;
  acquisitionUnitPrice: string | null;
  acquisitionQuantity: number | null;
  firstAssetAmount: string | null;
  firstAssetAccumAmount: string | null;
  acquisitionAmount: string | null;
  acquisitionAccumAmount: string | null;
  inquiryPointAcquisitionAccumAmount: string | null;
  inquiryPointBookValue: string | null;
  structureNameByType: string | null;
  usefulLifeYears: number | null;
  depreciationMethod: string | null;
  depreciationAccumAmount: string | null;
  conversionDepreciationAccumAmount: string | null;
  capitalExpenditureYn: string | null;
  leaseAssetYn: string | null;
  investmentCenterCode: string | null;
  businessAreaCode: string | null;
  linkageType: string | null;
  expenditureAccountingType: string | null;
  appliedYn: string | null;
  rejectionReason: string | null;
  voucherNo: string | null;
  expenditureSummary: string | null;
  causeActNo: string | null;
  expenditureRegDate: Date | null;
  expenditureApproveDate: Date | null;
  expenditureResolutionAmount: string | null;
  paymentOrderNo: string | null;
  paymentOrderDate: Date | null;
  propertyCode: string | null;
  ownershipTypeCode: string | null;
  propertyUseTypeCode: string | null;
  administrativePropertyTypeCode: string | null;
  propertyManagerCode: string | null;
  deputyManagerCode: string | null;
  delegatedManagerCode: string | null;
  propertyValue: string | null;
  propertyAcquisitionDate: Date | null;
  propertyAcquisitionAmount: string | null;
  acquisitionMethod: string | null;
  acquisitionReason: string | null;
  propertyCount: number | null;
  propertyArea: string | null;
  oldAssetName: string | null;
  changeReason: string | null;
  expenditureInOutType: string | null;
  disposalDate: Date | null;
  disposalAmount: string | null;
  disposalVoucherNo: string | null;
  landLotCode: string | null;
  actualLotCode: string | null;
  area: string | null;
  actualArea: string | null;
  acquisitionArea: string | null;
  sharedLandShare: string | null;
}